Crate depile

depile

“de-” for “apart”, as opposed to “com-” for “together”.

Translate three-address code back to C code.

In principle, one should be able to find specification of the input format on this lab description page. However, there is no formal definition (in BNF, EBNF, or whichever variant of BNF at your choice, as one would usually expect); only some informal discussion can be found in Section “The 3-Address Intermediate Format”.

Fortunately, this three-address code format is used nowhere else (and we are, and hopefully will be, the only ones to suffer from poor documentation). Since no other tool than csc (bundled in lab materials) should ever be producing files in this format, we assume that the input will always appear as if it were generated by csc, with a few relaxations.

Did You Know?: year 2022 has arrived, yet the whole build process of csc is managed not even with Makefile (!), but with a minimalistic Bash script assuming the local C compiler always being gcc…

“Take this, noob! This is how we program in the REAL WORLD!”

License

This project is licensed under the GNU Affero General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.

Input Specification

Input is handled per line. Lines are trimmed, and empty lines are filtered out.

Every line starts with a header indicating the instruction index, followed by the instruction:

instr 14: add res_base#32744 GP

Instructions are parsed into ir::Instrs structurally.

Re-exports

pub use cli::Cli;

Modules

analysis

Various analyses.

cli

Command line interface support.

ir

IR: Intermediate representations.