Trait depile::analysis::control_flow::ControlFlow

pub trait ControlFlow {
    fn entry_block_idx(&self) -> usize;

    fn block_count(&self) -> usize;

    fn successor_blocks(&self, block_idx: usize) -> NextBlocks;

    fn collect_reachable_into(
        &self, 
        block_idx: usize, 
        result: &mut BTreeSet<usize>
    ) { ... }

    fn collect_reachable(&self, block_idx: usize) -> Vec<usize> { ... }
}

A control flow: a series of basic blocks indexed 0..block_count(), with a successor relation given by successor_blocks. Some helper functions are also provided here.

Required methods

fn entry_block_idx(&self) -> usize

Index of the entry block.

fn block_count(&self) -> usize

Get the total number of basic blocks in this control flow.

fn successor_blocks(&self, block_idx: usize) -> NextBlocks

Which blocks are following this one (in the control flow graph)?

Provided methods

fn collect_reachable_into(&self, block_idx: usize, result: &mut BTreeSet<usize>)

Collect all the blocks reachable from the given block into an existing set.

fn collect_reachable(&self, block_idx: usize) -> Vec<usize>

Calculate all the blocks reachable from the given block.