Trait depile::analysis::control_flow::ControlFlowExt

pub trait ControlFlowExt: ControlFlow {
    type BlockKind: InstrExt;
    fn get_block(&self, block_idx: usize) -> &Block<Self::BlockKind>;

    fn entry_block(&self) -> &Block<Self::BlockKind> { ... }
}

More control flow related API.

Associated Types

type BlockKind: InstrExt

The kind of the blocks in this control flow.

Required methods

fn get_block(&self, block_idx: usize) -> &Block<Self::BlockKind>

Get the block content for a given index.

Provided methods

fn entry_block(&self) -> &Block<Self::BlockKind>

Get the block content for the entry block, equivalent to get_block on ControlFlow::entry_block_idx.