Module depile::ir::instr

Instructions for three-address code.

Modules

basic

Basic instructions for “raw” three-address code, with 1-1 correspondence to valid input texts.

resolved

Instructions, with inter-procedural instructions resolved, i.e. basic::InterProc::PushParam and basic::InterProc::Call are merged into a single instruction InterProc::Call.

stripped

Stripped instructions for use in Functions.

structured

“Structured” instructions. In this flavour of instructions, branching instructions own their successor blocks. Raw gotos are eliminated, and all of them are turned into control flow structures.

Structs

Branching

Common branching instructions.

Kind

An instruction kind.

Enums

BOp

Binary operators.

BranchKind

Branching methods.

Instr

Instructions.

Never

Replacement for the not yet stable “never” type.

UOp

Unary operators.

Traits

HasDest

Manipulate the “destination” field in an instruction.

HasOperand

Number of operands for an instruction.

InstrExt

Extension point for instructions.

OutputInfo

Information on output of an instruction.