Module depile::ir::instr::basic

Basic instructions for “raw” three-address code, with 1-1 correspondence to valid input texts.

Also repurposed for block-level instructions. These are just like basic instructions, but with the exceptions that all destinations in Branching instructions are interpreted as block indices instead of instruction indices, and that the marker EntryPc shall never be used.

Enums

InterProc

Basic inter-procedural instructions.

Marker

Kind “basic” have some markups.

Operand

Operands to all the “basic” Instructions.

Type Definitions

Block

Block with kind “basic”.

Blocks

Blocks with kind “basic”.

BranchKind

BranchKind with kind “basic”.

Branching

Branching instructions with kind “basic”.

Extra

Kind “basic” have no extra instructions.

Function

Function with kind “basic”.

Functions

Functions with kind “basic”.

Instr

Instruction with kind “basic”, i.e. “raw” three-address code.

Kind

Instruction kind “basic”.