Trait depile::ir::instr::InstrExt

pub trait InstrExt: Copy + Eq + Ord + Debug + Display + FromStr {
    type Operand;
    type Branching;
    type Marker;
    type InterProc;
    type Extra;
}

Extension point for instructions.

The Self type is essentially insignificant, to avoid the awkwardness of manually specifying all the bounds using derivative, we instead impose the superfluous bounds here. Use Kind to have all the bounds automatically implemented.

Associated Types

type Operand

Operand type.

type Branching

Branching instructions, eventually replaced by structural control flow.

type Marker

Marker instructions, eventually completely eliminated.

type InterProc

Inter-procedural instruction, i.e. function call related stuff.

type Extra

Other instructions, e.g. φ-nodes in SSA form.